move platform code from bp_sim
