+++ linux/ws_main.py
@@ -125,6 +125,7 @@ main_src = SrcGroup(dir='src',
             'utl_mbuf.cpp',
              'inet_pton.cpp',
              'bp_sim.cpp',
+             'trex_platform.cpp',
              'bp_sim_stf.cpp',
              'utl_port_map.cpp',
              'os_time.cpp',
