+++ linux_dpdk/ws_main.py
@@ -250,6 +250,7 @@ main_src = SrcGroup(dir='src',
              'astf/astf_template_db.cpp',
              'stt_cp.cpp',
              'bp_sim.cpp',
+             'trex_platform.cpp',
              'utl_term_io.cpp',
              'utl_port_map.cpp',
              'global_io_mode.cpp',
