+++ src/trex_defs.h
@@ -23,6 +23,10 @@ limitations under the License.
 
 #define TREX_MAX_PORTS 16
 
+#define MAX_SOCKETS_SUPPORTED   (4)
+#define MAX_THREADS_SUPPORTED   (120)
+
+
 // maximum number of IP ID type flow stats we support. Must be in the form 2^x - 1
 #define MAX_FLOW_STATS 1023
 #define MAX_FLOW_STATS_X710 127
